+++ xmake/platforms/macosx/load.lua
@@ -68,7 +68,7 @@ function main()
     -- init flags for asm 
     local as = config.get("as")
     if as == "yasm" then
-        _g.asflags = { "-f", "macho64" }
+        _g.asflags = { "-f", arch == "x86_64" and "macho64" or "macho32" }
     else
         _g.asflags = { "-arch " .. arch }
         if xcode_sdkdir then